Introduce support for regex based groups (#55)
* Introduce support for regex based groups

Motivation:

Currently, only static groups are supported. It would be convenient to support regex based groups, eg for grouping `java` and `javax` packages in the same group.

Modifications:

* support the same syntax as scalafix-organize-imports
* Introduce Block type hierarchy
* Fix SortImports file name case

Result:

Regex based groups are now supported

* Add test proving regex overlapping behavior

* Improve WildcardAndGroupFirstSort, only give higher precedence after a dot

Motivation:

We're only interrested in give higher precedence to `_` and `{` when they're right after a `.`, not anywhere.

Modification:

Fix regexes.

Result:

Use standard ASCII sort for `_` that would be inside a package or class name.